// src/shared/api/generated/model/autoTransferResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * CoreBank API Specification
 * Corebank REST API 명세서
 * OpenAPI spec version: v1.0.0
 */
import type { AutoTransferResponseStatus } from './autoTransferResponseStatus';

export interface AutoTransferResponse {
  /** 자동이체 ID */
  autoTransferId?: number;
  /** 출금계좌 ID (내 계좌) */
  withdrawalAccountId?: number;
  /** 입금계좌번호 */
  depositAccountNumber?: string;
  /** 예금주명 */
  payeeName?: string;
  /** 회당 이체금액 */
  amount?: number;
  /** 이체주기(개월) */
  cycleMonths?: number;
  /** 이체지정일 */
  transferDay?: number;
  /** 이체 시작일 */
  startDate?: string;
  /** 이체 종료일 */
  endDate?: string;
  /** 다음 이체 예정일 */
  nextExecutionDate?: string;
  /** 내 통장 표시내용 */
  myPassbookMemo?: string;
  /** 상대 통장 표시내용 */
  recipientPassbookMemo?: string;
  /** 자동이체 상태 */
  status?: AutoTransferResponseStatus;
}